requirements.txt
flask==3.0.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2
joblib==1.4.2
inflect==7.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1
gunicorn==21.2.0
